~finish moving dt64 arith tests
